Extract the rewrite rules into Plumb::Optimizer
Plumb's operators do not build the tree you wrote: they build an equivalent one
with the provably-redundant runtime work removed. Those rewrite rules lived in
Plumb::Subtyping, mixed in with the subtype relation, which conflated two
different kinds of thing — the relation ANSWERS questions about types, the rules
REWRITE one AST into another.

Move 16 methods into lib/plumb/optimizer.rb: reduce_step and its attribute
narrowing, redundant_refinement? and the record case, reduce_union, factor_union
and the step flatten/rebuild helpers. Subtyping keeps the relation (subtype?,
atomic_subtype?), the memoized structural predicates (value_preserving?,
accepted_type, resolved_*) and #intersect — computing a greatest lower bound is
lattice algebra, not a rewrite. The dependency is now one-way: Optimizer ->
Subtyping, never back.

The rules are unchanged, and they still run EAGERLY from the operators rather
than as a deferred pass. That is deliberate: reductions are observable through
#inspect, #== and the visitors, so a type's identity is its reduced form, and
`(Integer[0..100] >> Integer[0..]) == Integer[0..100]` is a documented property.
All 518 lines of reduction_spec pass unchanged.

Three named entry points replace the inline rule chains in Composable, which
also names a distinction that was previously only implicit in which methods each
operator happened to call:

  - #rewrite_step        (#>>)  reduce, then absorption, then build
  - #rewrite_refinement  (#/)   reduce, then build — absorption is deliberately
                                skipped, because #/ exists to assert a narrowing
                                the checker cannot prove, so dropping it as
                                "already guaranteed" would discard the cast the
                                caller asked for
  - #rewrite_union       (#|)   absorb, then factor, then build

The module docstring lists the rules in the order they are tried and states the
four invariants every rewrite must preserve — validity, output value, errors and
execution order. Execution order is the one that is easy to lose and invisible
to most tests, since it only shows up when a step has a side effect; it is why
absorption and factoring are gated on #value_preserving?, and spec/
invariants_spec.rb asserts it with probes that log each step as it runs.

Performance measured before (08e7c25) and after across build, resolve, subtype
and json-schema workloads: unchanged within run-to-run noise.
