+++ lib/plumb.rb
@@ -139,6 +139,7 @@ require 'plumb/range_class'
 require 'plumb/interface_class'
 require 'plumb/attributes'
 require 'plumb/subtyping'
+require 'plumb/optimizer'
 require 'plumb/types'
 require 'plumb/codec'
 require 'plumb/json_schema_visitor'
